Added RepairAll

+++ EDDiscovery/EliteDangerous/JournalEvents/JournalPayLegacyFines.cs
@@ -13,9 +13,9 @@ namespace EDDiscovery.EliteDangerous.JournalEvents
         public JournalPayLegacyFines(JObject evt) : base(evt, JournalTypeEnum.PayLegacyFines)
         {
             Amount = Tools.GetInt(evt["Amount"]);
-            BrokerPercentage = Tools.GetInt(evt["BrokerPercentage"]);
+            BrokerPercentage = Tools.GetDouble(evt["BrokerPercentage"]);
         }
         public int Amount { get; set; }
-        public int BrokerPercentage { get; set; }
+        public double BrokerPercentage { get; set; }
     }
 }
