+++ EDDiscovery/EliteDangerous/JournalEvents/JournalRedeemVoucher.cs
@@ -15,11 +15,11 @@ namespace EDDiscovery.EliteDangerous.JournalEvents
         {
             Type = Tools.GetStringDef(evt["Type"]);   
             Amount = Tools.GetInt(evt["Amount"]);
-            BrokerPercentage = Tools.GetInt(evt["BrokerPercentage"]);
+            BrokerPercentage = Tools.GetDouble(evt["BrokerPercentage"]);
         }
         public string Type { get; set; }
         public int Amount { get; set; }
-        public int BrokerPercentage { get; set; }
+        public double BrokerPercentage { get; set; }
     }
 }
